Adding TTL to proxied messages
A message proxied to another node can be proxied again.
This is clearly evident when picking `Horde.UniformRandomDistribution` as a
distribution strategy, especially in combination with a large number of nodes.

This commit contains an implementaiton which limits the amount of times a message
can be proxied before expiring, by adding a TTL which works similar to the TTL
on IP packets.